ENH: moved SliceTrackerOverviewStep into separate file
* onZFrameRegistration successful, session invokes event and OverviewStep is activated
